pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "migration-scraper"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.7",
]

[project.scripts]
migration-scraper = "scraper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["scraper"]
